# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "집결 서버",
    "version": "0.1.0"
  },
  "paths": {
...
    "/auth/kakao/url": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Kakao Url",
        "operationId": "kakao_url_auth_kakao_url_get",
        "parameters": [
          {
            "name": "redirectUri",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 500,
              "title": "Redirecturi"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthUrlResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/kakao": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Kakao Login",
        "operationId": "kakao_login_auth_kakao_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "502": {
            "description": "Bad Gateway",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/me/profile": {
      "get": {
        "tags": [
          "me"
        ],
        "summary": "Read Profile",
        "operationId": "read_profile_me_profile_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Profile"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "me"
        ],
        "summary": "Write Profile",
        "operationId": "write_profile_me_profile_put",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Profile"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Profile"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AuthRequest": {
        "properties": {
          "code": {
            "type": "string",
            "minLength": 1,
            "title": "Code"
          },
          "redirectUri": {
            "type": "string",
            "minLength": 1,
            "title": "Redirecturi"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "redirectUri"
        ],
        "title": "AuthRequest"
      },
      "AuthResponse": {
        "properties": {
          "token": {
            "type": "string",
            "title": "Token"
          },
          "user": {
            "$ref": "#/components/schemas/AuthUser"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "token",
          "user"
        ],
        "title": "AuthResponse"
      },
      "AuthUrlResponse": {
        "properties": {
          "url": {
            "type": "string",
            "title": "Url"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "url"
        ],
        "title": "AuthUrlResponse"
      },
      "AuthUser": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "nickname": {
            "type": "string",
            "title": "Nickname"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "nickname"
        ],
        "title": "AuthUser"
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Profile": {
        "properties": {
          "household": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProfileHousehold"
              },
              {
                "type": "null"
              }
            ]
          },
          "address": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProfileAddress"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "Profile"
      },
      "ProfileAddress": {
        "properties": {
          "road": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Road"
          },
          "jibun": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Jibun"
          },
          "sidoCd": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^\\d{2}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sidocd"
          },
          "sigunguCd": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^\\d{5}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sigungucd"
          },
          "bjdongCd": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^\\d{5}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bjdongcd"
          },
          "platGbCd": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[01]$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Platgbcd"
          },
          "bun": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^\\d{4}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bun"
          },
          "ji": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^\\d{4}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ji"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ProfileAddress",
        "description": "주소 기본값. 채워진 것만 담는 부분 객체."
      },
      "ProfileHousehold": {
        "properties": {
          "size": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 20.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Size"
          },
          "income": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "le48",
                  "48_60",
                  "60_100",
                  "gt100",
                  "unknown"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Income"
          },
          "housingBenefit": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "yes",
                  "no",
                  "unknown"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Housingbenefit"
          },
          "tenure": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "own",
                  "rent",
                  "public_rent",
                  "unknown"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Tenure"
          },
          "traits": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "enum": [
                    "elderly65",
                    "disabled",
                    "welfare"
                  ]
                },
                "type": "array"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Traits"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ProfileHousehold",
        "description": "가구 기본값. 채워진 것만 담는 부분 객체."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}